:art: reformat filter tags and fix some css
